# templates/schemas/api-contract.yaml
openapi: 3.0.3
info:
  title: FutureReady API
  version: 1.0.0
  description: AI-powered full-stack code generation platform
servers:
  - url: https://api.futureready.dev/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local Development

paths:
  /health:
    get:
      summary: Health check
      responses:
        200:
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
                  version: { type: string, example: "1.0.0" }

  /auth/register:
    post:
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        201:
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        400:
          description: Email already registered

  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        401:
          description: Invalid credentials

  /auth/me:
    get:
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        200:
          description: Current user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'

  /projects:
    get:
      summary: List projects
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    post:
      summary: Create project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
      responses:
        201:
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'

  /projects/{project_id}:
    get:
      summary: Get project by ID
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        404:
          description: Project not found

  /reports/generate:
    post:
      summary: Generate report
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                project_id: { type: string, format: uuid }
                format: { type: string, enum: [pdf, json] }
      responses:
        200:
          description: Report queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'

components:
  schemas:
    UserCreate:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        full_name: { type: string }

    UserOut:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        full_name: { type: string }
        is_active: { type: boolean }
        role: { type: string }

    Token:
      type: object
      properties:
        access_token: { type: string }
        token_type: { type: string, example: bearer }

    Project:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        status: { type: string, enum: [draft, in_progress, review, completed, archived] }
        owner_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    Report:
      type: object
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        format: { type: string, enum: [pdf, json] }
        status: { type: string, enum: [queued, processing, completed, failed] }
        download_url: { type: string, format: uri }

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT